Refactor mod `drivers`: Extract out sub mods
